% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorReg2D.R
\name{\%i\%}
\alias{\%i\%}
\title{Computation of the vector of inner product values of the matrices
\code{X[,,i]} and \code{B}.}
\usage{
X \%i\% B
}
\arguments{
\item{X}{A numerical 3D array. Each slice is a matrix of size the same as \code{B}.}

\item{B}{A numerical matrix.}
}
\value{
numerical vector.
}
\description{
Computation of the vector of inner product values of the matrices
\code{X[,,i]} and \code{B}.
}
